refactor(cketh): fold the sweep-spending tests into four
Reviewing the first draft as Gregory would have:

Two of the ten tests were the same test — identical setup, call and
assertion, differing only in name and doc — and the five that called the
accounting directly asserted a subset of what the five event-driven ones
already covered. His standard on this stack was to drive the production path,
so the direct-call set goes and the event-driven set stays, moved to the
accounting's own test file where its subject lives.

The three outcome cases — a sweep that succeeds carrying no value, one that
succeeds carrying value, one that fails — differ only in what the sweeper
ends up paying, so they are one table now, and it reads as the refund rule
itself. The failure row names the outcome and the value in its assertion, so
a failing row still says which one it was.

Two fixtures stop being copies: the sweep request moves to `test_fixtures`
beside the funding one, and the sweep tests arrange their delivered ETH
directly rather than borrowing the funding task's helper across modules.

The arithmetic moves off the audit arms and onto `State`, next to
`update_balance_upon_withdrawal`, which is where the withdrawal pipeline
keeps the same work, and the fee-ceiling `expect` now says why it holds.

Ten tests become four, and both wirings still fail under mutation: dropping
the provision fails two of them, dropping the refund fails the table.
