+++ rs/ethereum/cketh/minter/src/sweeper/tests.rs
@@ -118,7 +118,7 @@ mod concurrent_fundings {
     }
 
     /// Accepts a funding and runs it to a finalized transfer.
-    pub(super) fn fund(state: &mut State, request: &EthWithdrawalRequest) {
+    fn fund(state: &mut State, request: &EthWithdrawalRequest) {
         accept(state, request);
         let transaction = create(state, request);
         let signed = sign(state, request, transaction);
@@ -283,214 +283,5 @@ mod concurrent_fundings {
                 .is_none(),
             "with both settled, nothing is outstanding and the guard holds nothing back"
         );
-    }
-}
-
-/// The other side of the balance bound: sweeping spends the gas a funding provisioned, and the bound
-/// has to come back down as it does, or the minter believes a drained sweeper is still topped up and
-/// never funds it again.
-mod sweep_spending {
-    use crate::eth_rpc_client::responses::{TransactionReceipt, TransactionStatus};
-    use crate::lifecycle::EthereumNetwork;
-    use crate::numeric::{BlockNumber, GasAmount, LedgerBurnIndex, Wei, WeiPerGas};
-    use crate::state::State;
-    use crate::state::audit::{EventType, apply_state_transition};
-    use crate::state::tests::eth_balance_of;
-    use crate::state::transactions::{PipelineRequest, SweepId, SweepRequest};
-    use crate::sweep::SWEEP_TRANSACTION_GAS_LIMIT;
-    use crate::test_fixtures::{initial_state, sweeper_funding_request};
-    use crate::tx::{GasFeeEstimate, SignedSweepTransaction, TransactionSignature};
-    use ic_ethereum_types::Address;
-
-    const MINIMUM_BURN: u128 = 30_000_000_000_000_000; // 0.03 ETH, ckETH's mainnet minimum
-    const TARGET: u128 = 10 * MINIMUM_BURN;
-    /// What the receipt below charges: the sweep's gas limit at one wei per gas.
-    const SWEEP_GAS: u64 = 100_000;
-
-    fn state() -> State {
-        let mut state = initial_state();
-        state.cketh_minimum_withdrawal_amount = Wei::new(MINIMUM_BURN);
-        state.eth_balance = eth_balance_of(Wei::new(10 * TARGET));
-        state
-    }
-
-    /// Puts ETH at the sweeper address the way production does: a funding, run to a finalized
-    /// transfer, which is what credits the bound in the first place.
-    fn deliver_to_sweeper(state: &mut State) -> Wei {
-        let request = crate::state::transactions::EthWithdrawalRequest {
-            ledger_burn_index: LedgerBurnIndex::new(1),
-            ..sweeper_funding_request(Wei::new(TARGET))
-        };
-        super::concurrent_fundings::fund(state, &request);
-        state.sweeper_funding.sweeper_balance_lower_bound()
-    }
-
-    /// `amount` is the ETH the transaction itself moves: zero for an ERC-20 sweep, which moves
-    /// tokens through the calldata, and non-zero only where the call needs value.
-    fn sweep_request(amount: Wei) -> SweepRequest {
-        SweepRequest {
-            id: SweepId(1),
-            destination: Address::new([0x77; 20]),
-            amount,
-            data: vec![0xaa],
-            max_transaction_fee: Wei::from(SWEEP_FEE_CEILING),
-            created_at: 1_620_328_630_000_000_000,
-            authorizations: vec![],
-        }
-    }
-
-    /// Accepts a sweep moving `amount` of ETH, which is where its provision is taken.
-    fn accept_sweep(state: &mut State, amount: Wei) -> SweepRequest {
-        let request = sweep_request(amount);
-        apply_state_transition(state, &EventType::AcceptedSweepRequest(request.clone()));
-        request
-    }
-
-    /// Runs an accepted sweep through the pipeline to a finalized receipt with `status`.
-    fn finalize_sweep(state: &mut State, request: &SweepRequest, status: TransactionStatus) {
-        let request = request.clone();
-        let sweep_id = request.id;
-
-        let transaction = request
-            .create_transaction(
-                state.sweeper_transactions.next_transaction_nonce(),
-                GasFeeEstimate {
-                    base_fee_per_gas: WeiPerGas::ONE,
-                    max_priority_fee_per_gas: WeiPerGas::ONE,
-                },
-                SWEEP_TRANSACTION_GAS_LIMIT,
-                EthereumNetwork::Mainnet,
-            )
-            .expect("test setup: the sweep must cover its transaction fee");
-        apply_state_transition(
-            state,
-            &EventType::CreatedSweeperTransaction {
-                sweep_id,
-                transaction: transaction.clone(),
-            },
-        );
-
-        let signed = SignedSweepTransaction::from((
-            transaction,
-            TransactionSignature {
-                signature_y_parity: false,
-                r: Default::default(),
-                s: Default::default(),
-            },
-        ));
-        apply_state_transition(
-            state,
-            &EventType::SignedSweeperTransaction {
-                sweep_id,
-                transaction: signed.clone(),
-            },
-        );
-        apply_state_transition(
-            state,
-            &EventType::FinalizedSweeperTransaction {
-                sweep_id,
-                transaction_receipt: TransactionReceipt {
-                    block_hash:
-                        "0xce67a85c9fb8bc50213815c32814c159fd75160acf7cb8631e8e7b7cf7f1d472"
-                            .parse()
-                            .unwrap(),
-                    block_number: BlockNumber::new(4190269),
-                    effective_gas_price: WeiPerGas::ONE,
-                    gas_used: GasAmount::from(SWEEP_GAS),
-                    status,
-                    transaction_hash: signed.hash(),
-                },
-            },
-        );
-    }
-
-    /// The fee ceiling the fixture's sweep provisions against.
-    const SWEEP_FEE_CEILING: u64 = 10 * SWEEP_GAS;
-
-    #[test]
-    fn should_provision_an_accepted_sweep_out_of_the_bound() {
-        let mut state = state();
-        let delivered = deliver_to_sweeper(&mut state);
-
-        accept_sweep(&mut state, Wei::ZERO);
-
-        assert_eq!(
-            state.sweeper_funding.sweeper_balance_lower_bound(),
-            delivered.checked_sub(Wei::from(SWEEP_FEE_CEILING)).unwrap(),
-            "the bound must fall by the most the sweep can cost, before it has cost anything"
-        );
-    }
-
-    #[test]
-    fn should_settle_the_bound_on_what_a_sweep_actually_paid() {
-        let mut state = state();
-        let delivered = deliver_to_sweeper(&mut state);
-        let request = accept_sweep(&mut state, Wei::ZERO);
-
-        finalize_sweep(&mut state, &request, TransactionStatus::Success);
-
-        assert_eq!(
-            state.sweeper_funding.sweeper_balance_lower_bound(),
-            delivered.checked_sub(Wei::from(SWEEP_GAS)).unwrap(),
-            "the unspent part of the fee ceiling comes back"
-        );
-    }
-
-    #[test]
-    fn should_hand_back_the_value_a_failed_sweep_did_not_move() {
-        let mut state = state();
-        let delivered = deliver_to_sweeper(&mut state);
-        // A sweep whose call carries ETH, so that the value is a separate thing to hand back.
-        let value = Wei::from(7 * SWEEP_GAS);
-        let request = accept_sweep(&mut state, value);
-        assert_eq!(
-            state.sweeper_funding.sweeper_balance_lower_bound(),
-            delivered
-                .checked_sub(value.checked_add(Wei::from(SWEEP_FEE_CEILING)).unwrap())
-                .unwrap(),
-            "test setup: the provision covers the value as well as the fee ceiling"
-        );
-
-        finalize_sweep(&mut state, &request, TransactionStatus::Failure);
-
-        assert_eq!(
-            state.sweeper_funding.sweeper_balance_lower_bound(),
-            delivered.checked_sub(Wei::from(SWEEP_GAS)).unwrap(),
-            "a failed sweep paid its gas and moved nothing, so the value comes back"
-        );
-    }
-
-    #[test]
-    fn should_keep_the_value_a_successful_sweep_moved() {
-        let mut state = state();
-        let delivered = deliver_to_sweeper(&mut state);
-        let value = Wei::from(7 * SWEEP_GAS);
-        let request = accept_sweep(&mut state, value);
-
-        finalize_sweep(&mut state, &request, TransactionStatus::Success);
-
-        assert_eq!(
-            state.sweeper_funding.sweeper_balance_lower_bound(),
-            delivered
-                .checked_sub(value.checked_add(Wei::from(SWEEP_GAS)).unwrap())
-                .unwrap(),
-            "the value left the sweeper along with the gas it paid"
-        );
-    }
-
-    /// Sweeping must not disturb the burn-first invariant: this ETH was counted as spent once
-    /// already, when the funding that delivered it finalized.
-    #[test]
-    fn should_leave_the_burn_first_accounting_alone() {
-        let mut state = state();
-        deliver_to_sweeper(&mut state);
-        let burned = state.sweeper_funding.cumulative_burned();
-        let spent = state.sweeper_funding.cumulative_spent();
-        let request = accept_sweep(&mut state, Wei::ZERO);
-
-        finalize_sweep(&mut state, &request, TransactionStatus::Success);
-
-        assert_eq!(state.sweeper_funding.cumulative_burned(), burned);
-        assert_eq!(state.sweeper_funding.cumulative_spent(), spent);
     }
 }
